+++ src/main/java/com/example/cms/page/PageRepository.java
@@ -1,12 +1,7 @@
-package com.example.cms.pages;
+package com.example.cms.page;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface PageRepository extends JpaRepository<Page, Long> {
-    List<Page> findAllByParent(Page parent);
-
     boolean existsByParent(Page parent);
-
 }
