+++ src/main/java/com/example/cms/page/exceptions/PageException.java
@@ -1,4 +1,4 @@
-package com.example.cms.pages.exceptions;
+package com.example.cms.page.exceptions;
 
 
 import com.example.cms.validation.exceptions.BadRequestException;
@@ -11,15 +11,15 @@ public class PageException extends BadRequestException {
 
     private static String getMessage(PageExceptionType type) {
         switch (type) {
-            case NotFoundParent:
+            case NOT_FOUND_PARENT:
                 return "Parent was not found";
-            case NotFoundUniversity:
+            case NOT_FOUND_UNIVERSITY:
                 return "University was not found";
-            case NotFoundUser:
+            case NOT_FOUND_USER:
                 return "User was not found";
-            case DeletingPageWitchChild:
+            case DELETING_WITH_CHILD:
                 return "Cannot delete page with at least one child";
-            case IdEqualsParentId:
+            case ID_SAME_AS_PARENT:
                 return "Page id must be different than parent page id";
         }
         return "Unknown bad request";
