+++ src/main/java/com/example/cms/page/exceptions/PageExceptionType.java
@@ -1,0 +1,10 @@
+package com.example.cms.page.exceptions;
+
+public enum PageExceptionType {
+    NOT_FOUND_PARENT,
+    NOT_FOUND_UNIVERSITY,
+    NOT_FOUND_USER,
+    DELETING_WITH_CHILD,
+    ID_SAME_AS_PARENT
+}
+
