--- src/main/java/com/example/cms/pages/exceptions/PageExceptionType.java
@@ -1,10 +1,0 @@
-package com.example.cms.pages.exceptions;
-
-public enum PageExceptionType {
-    NotFoundParent,
-    NotFoundUniversity,
-    NotFoundUser,
-    DeletingPageWitchChild,
-    IdEqualsParentId
-}
-
