+++ src/main/java/com/example/cms/university/UniversityRepository.java
@@ -1,0 +1,14 @@
+package com.example.cms.university;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+public interface UniversityRepository extends JpaRepository<University, Long> {
+
+    @Query("SELECT university FROM University university WHERE university.name = ?1")
+    Optional<University> findUniversitiesByName(String name);
+}
