+++ src/main/java/com/example/cms/user/UserRepository.java
@@ -2,8 +2,6 @@ package com.example.cms.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.List;
-
 public interface UserRepository extends JpaRepository<User, Long> {
 
 }
