+++ src/main/resources/application.yml
@@ -14,7 +14,3 @@ spring:
   jpa:
     hibernate:
       ddl-auto: create
-
-server:
-    error:
-      include-message=always:
